+++ pyproject.toml
@@ -235,6 +235,7 @@ omit = [
     "src/maps/*",
     "src/api/api_data_fetcher.py",
     "src/export/const_definitions_exporter.py",
+    "src/export/license_export_utils.py",
     "src/export/org_admin_exporter.py",
     "src/export/org_alarm_event_exporter.py",
     "src/export/org_template_exporter.py",
